[ADDED] KV: ability to delete purges older than a certain time
A new option when creating the KeyValueStore allow specifying
the threshold that dictates if the purge/delete marker would be
removed or not during a PurgeDeletes() call. By default, this
threshold is set to 30 minutes, that is, markers that are older
than that would be removed, the ones that are younger are kept.
The data is always removed.

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>